+++ crates/nu-protocol/src/engine/mod.rs
@@ -5,6 +5,9 @@ mod engine_state;
 mod overlay;
 mod pattern_match;
 mod stack;
+mod state_delta;
+mod state_working_set;
+mod usage;
 
 pub use call_info::*;
 pub use capture_block::*;
@@ -13,3 +16,5 @@ pub use engine_state::*;
 pub use overlay::*;
 pub use pattern_match::*;
 pub use stack::*;
+pub use state_delta::*;
+pub use state_working_set::*;
